Move kbuild parameters into a struct
Soong will use a new module type, which will generate the custom
Blueprint rules needed for kernel module builds.

To pass the relevant information from the "original" type to the build
type, allow the output of `generateKbuildArgs` to be stored in a struct,
rather than a dictionary. This will allow it to be embedded in the new
module type.

Change-Id: I4165a574ed6a5b297aa50055b63ac64b88f76a93
Signed-off-by: Chris Diamand <chris.diamand@arm.com>