Soong kernel module
Implement skeleton support for kernel modules in Soong. This is done by
creating an instance of a new "backend" module type, which can then
generate the required custom Ninja rules.

The generated rules can be seen in `build.ninja`; however, they are not
yet complete or correct:
 * Other modules' `Module.symvers` files are ignored
 * Relative kernel dirs are not handled
 * The rule is not executed when running `mm`
 * The installation path is hardcoded

These will be fixed in subsequent commits.

Change-Id: I9ce21401170d23268b4f2b51b882af465099269e
Signed-off-by: Chris Diamand <chris.diamand@arm.com>