Improve clang configurability
Introduce a set of options that make Bob pass argument to Clang so
that it references the GNU toolchain for certain items.

Modify toolchain.go so that we only set -L, -B and -isystem if one of
these options is set. Since we can now protect them with relevant
options, move the addition of libgcc library location and STL header
paths to newToolchainClangCommon. The usage of useLibgcc and useGnuStl
are now in a single function, so remove them from the Clang toolchain
structure.

Change-Id: I9fb203cec0a7fb84bb21abfb9d1e2f53ea3eb4e7
Signed-off-by: David Kilroy <david.kilroy@arm.com>