Get BuildDir from command line
Blueprint's `bootstrap` module already gets the build directory when it
parses the command line. This means we can use `bootstrap.BuildDir`,
instead of redundantly reading it from the environment.

However, this approach will not work on Soong. Accordingly, `builddir`
is moved into a helper in `standalone.go`, which will make it obvious,
via a compilation error, if we try and use it uninitialised in Soong.

Note that this requires some tweaks to how jsonPath and envHash are
defined and used. Primarily their main use is in the
dependencySingleton. This is simplified to a single call to
`pctx.AddNinjaFileDeps` from the Main() function, which means those two
globals can be removed.

Add this call in Soong too, which fixes the lack of equivalent
functionality in Soong.

Change-Id: I6eb766904efafed449aad99a3210085be50ed01d
Signed-off-by: Chris Diamand <chris.diamand@arm.com>