Don't use Soong's genrule for bob_generated_source
Implement a custom backend module type to handle generated sources,
libraries, and binaries. This will allow more flexible handling of
sources vs headers, which Soong's `genrule` does not allow.

Transformed sources will be converted to use the new backend module type
in a subsequent commit.

For now, the `${xxmod_out}` variable is not implemented, but is
explicitly set to the empty string. This lets the `generate_source` test
`build.bp` files parse so that the other tests can run.

Change-Id: I7ee83eebd7d64debc0f827bc910c78ed713f24cd
Signed-off-by: Chris Diamand <chris.diamand@arm.com>