Use custom backend for transformed sources
Allow transformed source modules to use the `genBackend` module. This is
done by passing the `Out` properties straight to `genBackend`, and using
the `inoutForSrc()` helper (refactored in b21b028) to generate the list
of inouts.

Note that this commit is not enough to enable the transform_source
tests, because ModuleDir() on Soong includes more path elements,
specifically the path from the Android root to the project dir. This
breaks `bob_transform_source` modules' assumptions about where their
generated header files will end up.

This will be fixed, and the tests enabled, in a subsequent commit.

Change-Id: I492be28e2160335fd28db979c0cd848e34724799
Signed-off-by: Chris Diamand <chris.diamand@arm.com>