Refactor inout helper
Factor out the inside of transformSource's `Inouts` loops so that it can
be re-used from the Soong plugin.

This commit is modified from the previous attempt to commit this,
b21b028f56, so that ${g.bob.SrcDir} is not added to the output file
path as an infix. In inoutForSrc, the regular expression replace needs
to run on the srcs path before g.sourcePrefix() is added. In addition
we only want that prefix added for non-generated source inputs.

This is achieved by adding a filePath class, which allows us to pass a
single object around from which we can retrieve relevant path portions
for processing in different use cases.

Change-Id: I9ef92900623f7ebe297deeed5ae01a91bbad6bfb
Signed-off-by: David Kilroy <david.kilroy@arm.com>