Verify static libraries created from scratch
`ar` appends its input to the existing static library. Add modules
which change the name of the object they are compiling (they will
contain the same code) dependent on the configuration. Then in
build_tests.sh, change the configuration to ensure the static
libraries are created from scratch.

For duplicate symbols to be detected, we need to the static libraries
to be in shared libraries as whole static libraries.

Change-Id: I6ca161849f3abe0925b8882391bb01bcd7d5ee0e
Signed-off-by: David Kilroy <david.kilroy@arm.com>