Detect configuration changes on Android
Commit cf4f8a4c3c36 changed to use the bob package context to record
build.ninja's dependency on config.json (and the environment hash).

However for Android makefiles, since nothing in the bob package
context results in output to the build.ninja, these dependencies were
being dropped.

We could revert to recording the dependency in a SingletonType as
before. Instead, have the AndroidMkOrderer singleton write a
dummy build rule to the build.ninja.

Change-Id: I7095566aa20c05a63582795a37c03692d9dcb9df
Signed-off-by: David Kilroy <david.kilroy@arm.com>