Use getopts instead of getopt in Bob tests
When running Bob tests on OSX everything works correctly until
`tests/build_tests.sh` calls `tests/bootstrap.sh -o build-in-src`.

`tests/bootstrap.sh` invokes `getopt` to get the `BUILDDIR` argument
(that is supposed to be `build-in-src`), but this does not work and it defaults
to `BUILDDIR=build`, breaking everything that follows.

This is because `getopt` on OSX 10.13.3 is different from `getopt` on Linux.

This commit fixes this by switching to `getopts` which is dependent on the shell
and not on the OS.

Change-Id: I5035d30cfe8fcf05036bd5716875a732ad995858
Signed-off-by: Alexander Khabarov <alexander.khabarov@arm.com>