Fix static library test
The test added in `77f85ae Verify static libraries created from scratch`
missed a static library dependency, causing it to fail on Soong.

Add the missing dependency, as well as fixing some whitespace from the
same commit.

Add the `--no-undefined` linker option, too; the lack of this is why the
tests were able to pass on Linux.

Change-Id: I62e2e1bd3a75ad94379ca63c00ea6228b86ead6b
Signed-off-by: Chris Diamand <chris.diamand@arm.com>