Refactor config loading
Modify the loadConfig function so that it returns an error, and takes
the structure to load configuration into as an argument.

Eventually we want soong to load the configuration file, and we should
report errors nicely.

Also fix a format specifier in linux.go. %e is scientific notation for
floating point numbers.

Change-Id: I0dca554eef8c8c4c719246f0593ef463407ebc6b
Signed-off-by: David Kilroy <david.kilroy@arm.com>