Fix static_libs test on macOS
`ar` on macOS does not support `-N` flag and hence adding multiple
static libs with the same basename does not work. Further, `ld` on
macOS does not support `--no-undefined` option. This commit updates
static_libs test to address this. `--no-undefined` is disabled for `ld`
on macOS as its default behaviour is equivalent.

Change-Id: I2cb766ae9f987a9250b8641ead912aa771a47de4
Signed-off-by: Alexander Khabarov <alexander.khabarov@arm.com>