Actually link fake kernel modules together
Improve the kernel module testing by examining KBUILD_EXTRA_SYMBOLS to
try and actually link to related modules.

Add a dependency on $KDIR/Makefile - we should probably have this
anyway, but in this case it ensures that the kernel module tests will
still pass when doing incremental builds of the Bob tests.

Change-Id: I5b16dfe45f4785368e0f9b52b6d977ab484be1f4
Signed-off-by: Chris Diamand <chris.diamand@arm.com>