Use consistent Android make output directories
The stripped_binary tests were failing on Android make as the output
directories for shared libraries and binaries are inconsistent. Shared
libraries and binaries are being installed to vendor, but are not
marked as such.

Android Q detects the binaries are vendor as we are setting
LOCAL_MODULE_PATH_32/64 for them. We aren't setting these variables
for shared libraries (as they need to go where android expects them).
Just set the install groups to put these in non-vendor locations.

Change-Id: I3abd447fe34e90333319d8123c1b641c7c6cc813
Signed-off-by: David Kilroy <david.kilroy@arm.com>