Don't use parseAndAddVariationDeps in backend modules
Replace the calls to parseAndAddVariationDeps() in
genBackend.DepsMutator with native dependency adding functions. This is
because the former is for adding dependencies on "frontend" Bob modules,
complete with the Bob splitter variant. When adding dependencies between
backend modules, the variants used by the Bob splitter would not match.

For module_deps and module_srcs, use AddFarVariationDependencies() with
an empty variant list, which will automatically select the first variant
when multiple exist, which occurs with multilib builds where there are
32 and 64-bit versions of some modules.

Change-Id: I6e6e2ed59351cba0a7da94cea0a17bafc63b2356
Signed-off-by: Chris Diamand <chris.diamand@arm.com>