Further soong generate_libs fixups
Commit 9513398e6 introduced some issues, as it assumes that you can
cast to the genBackend type. You can't do this if the object is
actually genSharedLibraryBackend (or one of the others). Switch to use
interfaces in a number of cases.

When running on Android also setup the host clang path, and explicitly
use this for GEN_CC on soong.

Re-enable testing for generate_libs.

Change-Id: I0cea69af73d804a8f9645d29a83880761b6d5ab5
Signed-off-by: David Kilroy <david.kilroy@arm.com>