Always rename Bob modules
To support external libraries, we need to ensure that Bob's
`bob_external_[[shared|static]_library|binary]` declaration modules do
not conflict with the real modules, because, on the Soong plugin, Soong
and Bob modules share the same namespace.

Until now we avoided conflicts with the Bob and "backend" modules by
renaming the Bob modules just before generating the backends. However,
this will not work for external libraries, because the real modules are
parsed at the same time as the Bob modules, rather than being created
later.

Solve this by _always_ renaming the original Bob modules. This is
achieved by overriding `moduleBase().Name()` so that it returns as
suffixed version of the module name; i.e. the name is always the value
of the build.bp `name:` property, with a suffix.

This requires small changes throughout the code to ensure the new module
names are being used. Do this via `bobName()` and `buildbpName()`
helpers; the "Bob name" is the name including the suffix, whereas the
"build.bp name" is the user-visible name as specified in `build.bp`.

Change-Id: I5b5e3bd55e44f8f8345ac44f5852543e10d77abd
Signed-off-by: Chris Diamand <chris.diamand@arm.com>