Add initial Android.bp backend
Add bootstrap infrastructure for a new backend that will write
Android.bp files.

The test bootstrap script will bootstrap, configure and execute Bob
(to write the Android.bp file).

In the test project, define an install_path for all bob_install_groups
for the new backend. These are nominal paths at the moment, to prevent
the standard mutator from panicking.

Add compile time checks in the implementation of each generatorBackend
to ensure the appropriate functions are implemented. This avoids
relying on how the generators are used to idenitfy errors.

Also do a minor text cleanup in the test and example bootstrap scripts
that was noticed during review.

Change-Id: I06dfed5249d5196d6e5e668d15f0ab14a066686e
Signed-off-by: David Kilroy <david.kilroy@arm.com>