Define OS and BUILDER in a common file
The BUILDER_* configs need to be available in all projects, so we want
these to be picked up from a single file defined by Bob.

The default builder may depend on OS, so we include the OS definitions
as well. Only define the choice in Bob, and modify the Go code to
avoid checking the string config.

Fixup objcopy binary definitions to default based on toolchain rather
than OS.

Change-Id: Ied1e450b1c8078694f33cb542ce9065672cc0afd
Signed-off-by: David Kilroy <david.kilroy@arm.com>