Revert "Always rename Bob modules"
This reverts commit 79a8ed4907af90ac950cf7580694cdedd6923848 - the
Android.bp backend will run Bob in a separate process, so the names will
no longer conflict with the "real" modules built by Soong.

Disable the tests on the Soong plugin; it's more useful to revert this
earlier on, rather than keeping the Soong plugin 100% working, so we
don't get extra naming complexity creeping into the Android.bp
implementation.

Other than the above modification to `tests/bootstrap_soong`, this
commit is exactly as generated by `git revert`, and then including
everything except the copyright year changes. The comment at the end of
ccModuleName() is not removed, because it still applies.

Change-Id: I315a959dcfe1133f6903c4dd9a872a182a2b9fa1
Signed-off-by: Chris Diamand <chris.diamand@arm.com>