Show ldlib usage in graph
The graph produced by Bob can be useful for identifying what is linked
into a library. For this to be complete, we also need to show
libraries being linked via ldlibs. Only show the ldlib usage via
shared libraries and binaries (as the ldlibs property will exist on
static libraries, but not do anything).

Change-Id: I2ed29c5384e51a02a65b619bc8f94c28475574d6
Signed-off-by: David Kilroy <david.kilroy@arm.com>