Don't expand symlinks in relative_path
The Bob tests use a symlink to simulate embedding Bob in a project. The
bootstrap scripts then use various helpers, including `relative_path`,
to create symlinks from the build dir to the embedded "copy" of Bob.

`relative_path` currently uses `bob_realpath` on its arguments (since
`34736fc Fix bug in relative_path`), which means the `bob` symlink gets
expanded.

Until now, this hasn't mattered. However, the kernel module
Android.bp implementation will require that the `kmod_build.py` script
be at the same level as, or in a subdirectory of, the Android.bp file
using it. Because the `bob` symlinks is being expanded to `..`, this
will not work.

Fix this by using `abspath` in the `relative_path` implemention instead,
which will normalise the path without expanding symlinks.

Interestingly, this change also fixes the relative path tests on
merged-usr systems, where `/bin` is a symlink to `/usr/bin`.

Change-Id: I59cbf52a2eedfa201b6f5fa9736a0946818c2840
Signed-off-by: Chris Diamand <chris.diamand@arm.com>