relative_path: Expand source symlinks
The current implementation of `relative_path` in `pathtools.bash` does a
simple lexical calculation on the absolute paths to calculate the
result.

However, this is not correct if the source path contains a symlink. This
is because using `..` from inside a soft-linked directory will return
the parent directory of the *link target*, not the link location.

For example, consider the directory `/x/y`, and a symlink `/y -> /x/y`.

At first glance, and according to the current implementation,
the relative path from `/y` to somewhere outside it (say `/home`) should
be `../home`. However, because of the aforementioned reason, this is
incorrect. The `/y` symlink points  to `/x/y`, and therefore to get up
to the root directory we need *two* `../` path components. So the
correct relative path result from `/y` to `/home` is `../../home`.

The simple solution to this would be to simply expand all symlinks in
the source path. However, this doesn't really make sense without also
expanding the target, which we can't do, because of commit `dfcae5e
Don't expand symlinks in relative_path`. Consider for example a source
and target directory, both inside a symlinked dir (e.g. `/y/src` and
`/y/target`). The relative path from source to target is obviously
`../target`, regardless of the `/y` symlink; however, expanding just the
source symlink would result in comparing `/x/y/src` and `/y/target`,
resulting in `../../y/target` - a correct but unnecessarily complicated
answer.

Instead, limit symlink expanding to as few source path components as
possible. When we "chop off" a path component in order to find a common
prefix, we must expand links if that component is a symlink. If we do
not "reach" that path component, though, it is left as-is. This results
in the correct and simplest answer for both the "`/home` from `/y`"
*and* "`/y/target` from `/y/src`" cases.

Update the relative path tests to cover this. To handle the extra links
and directories created, simplify the cleanup code by running the tests
inside a temporary directory.

Remove the comment incorrectly stating that an absolute path will be
returned if it is shorter than the relative one.

Fix up some of the Bob scripts to use `cd -P` rather than `cd`; this
causes `cd` to go to the "physical" parent directory rather than the
logical, which is required now that paths inside `.bob.bootstrap` are
calculated w.r.t. the physical directory layout.

Change-Id: I4f16d64ead8e49466a7f2ea708fc85d64cadc481
Signed-off-by: Chris Diamand <chris.diamand@arm.com>