Change modules to record outputs in GenerateBuildActions
Previously Bob has inferred where all output can be found for
dependent modules based on the type of the module and its name. This
is nice as we can declare this up front, and use the outputs() and
outputDir() function anywhere. However, this is problematic for the
soong plugin.

Even though we are no longer sticking with the soong plugin, it makes
sense to retrieve the outputs from the module after its
GenerateBuildAction has run. This brings the code into alignment with
what the transform_source module has always done.

Create a simpleOutputProvider that modules can inherit. It implements
the outputDir(), outputs() and implicitOutputs() functions that will
be called to identify module outputs. It contains fields that need
to be populated when GenerateBuildActions is called to record the
outputs.

The getLibraryGeneratedPath() function is removed. Instead, call
m.outputs() to get the path where the generated library is output.

Outputs in the inout structure no longer include the output directory
when passed into the backend specific GenerateBuildAction code -
because this is the code needs to set the output directory. This also
affects depfile and rspfile in the inout structure.

For the generated libraries, the inouts() function is renamed to
generateLibraryInouts() and now takes the implicit outputs as an input
(it previously called outputs() and implicitOutputs()). Since some of
the functions are being removed, add interface checks on the gen_*
structures. The generateSource and transformSource Inouts() functions
are similary renamed to generateInouts()

The generatorBackend interface is removed from the following
functions, as it is no longer necessary:

 * outputDir()
 * outputs()
 * implicitOutputs()
 * filesToInstall()
 * getDepfile()
 * getRspfile()
 * getHeadersGenerated()
 * getSourcesGenerated()
 * getGeneratedFiles()

The filesToInstall() method just uses the outputs() function, which is
now implemented by the composed simpleOutputProducer. Therefore we can
eliminate some implementations in favour of methods on library and
generateLibrary.

Finally the generateBackend functions binaryOutputDir(),
staticLibOutputDir, sharedLibOutputDir() and kernelModOutputDir() are
removed. They are still defined and used in the Ninja and Android make
backends. The sourceOutputDir() function needs to stay for the moment,
as we need to disentangle some functionality with
export_gen_include_dirs.

Change-Id: I0e25a0be0cdedc078296c8bd745bd7c68f18cf67
Signed-off-by: David Kilroy <david.kilroy@arm.com>