Record generated include dirs in GenerateBuildActions
Add a helper type, headerProducer, that behaves like
simpleOutputProducer. Rather than tracking module outputs, it tracks
include directories. It implements the function genIncludeDirs() to
allow other modules to pick up the include directories that need to be
exported from the build output path.

Handle bob_generate_source encapsulation during the call to
GetGeneratedHeaders() during the GenerateBuildAction() rather than in
a mutator. This is because we will not have set the include
directories in the children at the point the mutator is run. Also this
function is only called if we need all the includes (and files) i.e.
not by the Android.bp backend.

In the Android.bp backend, we only need to forward the value of
Export_gen_include_dirs so we skip this processing.

This allows us to remove sourceOutputDir() from generatorBackend.

This fixes a few issues:

* The Android.bp backend was unnecessarily listing encapsulated
  include directories in Export_gen_include_dirs

* Bob was not listing encapsulated header files as order only
  dependencies.

Change-Id: Ia2bbd9f0f4e8980cdafeba0e5df5bfc1e87f6f64
Signed-off-by: David Kilroy <david.kilroy@arm.com>