Check for outdated Android.bp
Because the Android.bp file is generated separately to the actual build
(`mm` or equivalent), it is possible for build.bp files to have been
changed since running bootstrap_androidbp. This may cause incorrect
builds, e.g. if a build.bp was updated to adjust for a corresponding
change in the code.

Rather than continuing with the build and then ending up with a
compilation error, detect when this has occurred and output a warning.

This is done by creating a new genrule module, whose sources are the
build.bp files, and then checking that their hashes are the same as when
Android.bp was generated.

It would be possible to just implement this using mtimes. However:
 * The hashing is cheap, because we only do it when the build.bp files
   have been changed (i.e. when Bob is being run, or the genrule is
   being run because its source list contains something with a newer
   mtime).
 * Using a hash accounts is more correct; a warning is only printed when
   the build.bp files have _definitely_ changed. With an mtime check,
   the warning would be triggered if e.g. a change to a file was made,
   then reverted. This also means it is actually possible that we could
   *fail* the build when the warning is triggered, rather than just
   printing.

Reinstate the configProvider interface, to allow getConfig() to be
called from the SingletonContext when generating the new genrule.

Change-Id: Ia68cdbbf15bbd31eadc7965ad8186f2c71e9eac6
Signed-off-by: Chris Diamand <chris.diamand@arm.com>