Don't always calculate inouts
The GenerateBuildActions method on generated modules currently calls
`generateInouts` before invoking the backend-specific callback. Because
this is common code, it also gets invoked on Android.bp. However, on
Android.bp, it is incorrect to try and calculate the inouts, because in
the case of chained generator modules, only the backend plugin will know
the final paths. This means we cannot calculate an inout for each source
path, because we cannot know the location of generate source paths.

It is also obvious that this is the case, because the `inouts` argument
in androidBpGenerator is only used for the call to
`recordOutputsFromInout`, but the `outputs()` method is then never
used (and should not be, for the reasons described above).

Furthermore, trying to calculate the inouts is not harmless - it causes
a panic when using a `bob_generate_source` as an input to a
`bob_transform_source` via `module_srcs`:

  internal error: panic in GenerateBuildActions for module "transform_source_multiple_in"
  Path doesn't have as many elements as expected. f5.in

Move the calls to generateInouts into the backend methods. This fixes
the issue by allowing the inout generation to be omitted on Android.bp.

Change-Id: I1e83befa90e9503ea0654169826fe0a06cceff09
Signed-off-by: Chris Diamand <chris.diamand@arm.com>