Depend on `module_deps` outputs in transformed sources
Currently the genruleBob module calculates the implicit sources for a
module, including those required because of the `module_deps` field, but
then does _not_ pass them into the inouts created from Multi_out_srcs.

This means that bob_transform_source modules using module_deps may fail
to build, because Ninja is never informed of the dependency on the
dependent module's outputs. Note that it is possible for a build to
_appear_ correct, if by chance the dependent module is built first.

Fix this by including the common `implicits` list in the implicit
sources for transformed source inouts too.

Add a test which generates a source file to be used as a template for
each transformed source in a new transform_source module. This fits
quite neatly into the existing test pattern, as we just need to provide
an argument to use a specific file as the input template. However,
making this work sanely requires rewriting most of
`transform_source/generator.py` to be able to handle this.

Change-Id: I93811a82a638c8b3806980d670e02c3399893526
Signed-off-by: Chris Diamand <chris.diamand@arm.com>