Add support for testing the availability of compiler flags
Newer versions of a compiler adds support for new flags that would not be
understood by older compilers. To allow for the addition of those new flags
into the build.bp file, add '{{addIfSupported' late template
that will take as an argument the compiler flag to be tested for support
by the current toolchain compiler. Note that 'cflags' and 'export_cflags'
are tested with both the C and C++ compilers.

Signed-off-by: Liviu Dudau <liviu.dudau@arm.com>
Change-Id: Ifce31bb3d7df44d37ba0b16a72e8475980e5691e