Align generated source outputs on Android.bp
Soong's gen dirs are generally of the form `/path/to/module/gen`.
However, the Android.mk and Linux backends use the form
`build/gen/module_name`. Normally this doesn't matter, as everything is
contained within the gen dir, except when chaining multiple generated
modules. In this case, bob_transform_source used on Android.mk or Linux
may expect the module name to be included when doing the regex
replacement, and be exporting include directories accordingly.

This means that the transform_source tests are failing on Android.bp,
because the output path of `generate_source_to_transform` does not
contain the module name.

Fix this by adding a subdirectory named after the module inside Soong's
gen dir when calculating the outputs of generated modules.

Change-Id: I5e212943346288745d25bbd1b53c5e9f41d82ddc
Signed-off-by: Chris Diamand <chris.diamand@arm.com>
Signed-off-by: David Kilroy <david.kilroy@arm.com>