Fix match_srcs for Android.bp backend
match_srcs expects to find the matching sources relative to the
current module, but only one Android.bp gets written per project
and soong will always run from with TOP_DIR as the current directory.
As such, we need the full path for the source files to be returned
by the backend.

Signed-off-by: Liviu Dudau <liviu.dudau@arm.com>
Change-Id: I732f9a18ff103227c4e90c13a0f590b56665247d