Align flag quoting with soong
Flags should be written as the tool expects to see them when reading
its arguments. i.e. without any shell quoting.

Note that double quotes must be escaped for the blueprint parser.
Expansion of environment variables, ninja variables and make variables
is not supported in these properties.

This commit updates the handling of asflags, conlyflags, cflags and
cxxflags in the Linux and Android make backends.

Add tests for escaping in C and C++

Note that {{match_srcs}} expansions are expected to refer to Ninja
variables, so the properties are escaped prior to late template
expansion (taking care not to escape anything in the template).

This commit also cleans up some of the logic that sets the flags in
the different backends. In android_make hoist the ldflag calculation
out of the simialr if branches. In linux reorder cflags so that the
things we can escape are together (i.e. include flags now last)

Change-Id: Iff26320124214c852ecd87036de9fc78a3c57b7e
Signed-off-by: David Kilroy <david.kilroy@arm.com>