Test external shared libs with zlib
The existing external_lib tests are difficult to run outside of an
Android build, because it would be necessary to invoke another build
system somehow to build the test libraries.

To test this feature on Linux, add a simple test using the
commonly-available zlib library instead. This also requires adding the
missing options for pkg-config support to Mconfig.

`external_libs/build.bp` is also missing a copyright header - add one.

Change-Id: Id0cb02c6e3a6e9e546a66c2b9ec1758aa9198053
Signed-off-by: Chris Diamand <chris.diamand@arm.com>