Automatically split install paths
On Android, different parts of installation paths are handled
separately. Android.mk has separate LOCAL_MODULE_PATH and
LOCAL_MODULE_RELATIVE_PATH components, but on Soong, the module type and
a few extra attributes (e.g. `vendor`) determine where it ends up, and
the only extra control is to set a subdirectory within that location.

This means that we need to automatically translate a user-specified path
(e.g. `vendor/bin/testdir`) into "base" and "relative" components. In
this example, `vendor/bin`) is the "base", which on Soong, will be
determined by outputting a cc_binary module with `vendor: true`. The
rest of the path is the "relative" part, which is put into the
`relative_install_dir` property.

This is subtly different to the current behaviour, which simply copies
Bob's `relative_install_path` value into Soong's. This is unfortunately
incorrect, though, because some parts of the relative path may come from
an install group; currently, this would be ignored.

Change-Id: Idcfaca9c809893df84f61d914b6d4b41b3150c17
Signed-off-by: Chris Diamand <chris.diamand@arm.com>