Consolidate handling of test binaries
Use `tests` in the Android.bp backend as the install path designator
to install to a location for tests. The backend can then choose to
install to either testcases or data/nativetest depending on what is
supported. Initially this will need to install to data/nativetest as
we're not able to install Soong cc modules to testcases.

Use cc_test and cc_test_host to put binaries in data/nativetest. Turn
off cc_test default behaviors:

* Don't link to gtest

* Don't add Android.bp directory to search path

* Don't use directory per executable layout

* Don't generate a test configuration

Change-Id: I6ce3b209b0906ffa1a03dc45e9c4f76368301ade
Signed-off-by: David Kilroy <david.kilroy@arm.com>