Standardize Android.bp install path handling
We don't want the partition (system/vendor) to be in the install_path,
as this is expected to be inferred based on whether the vendor
property is set.

Therefore remove system and vendor from the Android.mk translations
and path splitting.

Fix the path handling to cope with $(TARGET_OUT_ETC)/firmware. Do
this by expanding Android make variables before splitting the path,
and translate compound paths.

Have binaries and shared libraries verify that their base install path
is set as expected.

Add and update documentation.

Change-Id: I8ba2f8a4e1d19a7f7c791895a4d4c28b4fe00690
Signed-off-by: David Kilroy <david.kilroy@arm.com>