Proprietary tests in data/nativetest
Add an accessor function to AndroidProps to identify if the module is
proprietary/vendor. This encapsulates the test on Owner.

Update all callers.

Update Android.bp backend resource handling to put tests in
data/nativetest/vendor if the module is proprietary. Setting the
vendor property doesn't change the behaviour of the prebuilt_data_bob
module

Fix the kernel install path when targetting `tests` - nativetest
should come before installRel. Also add vendor when the module is
proprietary.

Change-Id: I807bbef78c49512c8435817d87ac291eb8485c31
Signed-off-by: David Kilroy <david.kilroy@arm.com>