Allow tests to run with old gcc
__has_include is only supported from GCC 5.0. Where we use it, check
whether __has_include is defined before we use it. If it is not
available output a warning that we're not able to check whether a
header is available.

Change the command for gen_srcs_five to output a useful macro
definition to message.h, and actually use the header instead of using
__has_include.

Change-Id: I788b120e1eb8f8d34fa3f56058e5a0223846ba56
Signed-off-by: David Kilroy <david.kilroy@arm.com>