Rework strip.py arguments
In preparation for a similar script working with shared libraries,
standardize the options we want scripts to accept.

Primarily we want an option that identifies whether the library is
Unix-like Elf format, or OSX Mach-O. Use this format argument to
control the script behaviour instead of detecting a particular tool.

Add functions to get strip command flags in toolchain.go, and hook it
up to build.ninja code controlling library stripping.

Add definitions of dlsymutil and strip tools for use with the Xcode
toolchain.

Change-Id: I26400bea21b03d063f30f572fd02dded0fa0701b
Signed-off-by: David Kilroy <david.kilroy@arm.com>