Depend on shared libs via their toc file
Depend on toc files to avoid rebuilding callers when the API is
unchanged. When we do this we just need to change the implicit
dependencies on the shared libraries to an implicit dependency on the
toc file.

Notes:

* $in is unchanged, these are just compiled objects from srcs

* non compiled sources (things in srcs that are not .c/.cpp files) are
  kept as implicits. These are linker scripts or version scripts which
  still need to trigger a rebuild and relink.

* the shared library (.so) inputs get to the command line via ldlibs
  and don't need to change.

* we re-add the shared libraries as an order only dependency.
  Technically this shouldn't be needed as we have a transient
  dependency via the toc file. However if the user deletes the .so in
  the build directory we need the order on dependency to force the
  library to be rebuilt.

Allow disabling this functionality by setting the environment variable
BOB_ALWAYS_LINK_SHARED_LIBS=1. This allows people who encounter
problems with this mode of operation to work around it while the
majority of developers get the benefits automatically.

Change-Id: I72b9b92cb798632819068da6d213f9a7de0088d0
Signed-off-by: David Kilroy <david.kilroy@arm.com>