Clarify sharedLibrary outputFileName()
outputFileName() is expected to return the file name for a module's
output. Clarify that for shared libraries we always want this to be
the link library (i.e. the unversioned so) since this is what the
linker will be looking for.

When creating the build rules in sharedActions(), we can always use
getRealName() as this has the version appended if versioning is
enabled.

Change-Id: I019dc1525e63501d2392e8833cd650e288eb1bb9
Signed-off-by: David Kilroy <david.kilroy@arm.com>