Rework applying default properties to more generic way.
Currently every module which references `bob_defaults` has to be
equipped with the same set of properties no matter if it uses them
or not. This forces e.g. all library modules to contain kernel
properties.

Change the way of applying default properties to be more generic
and which allows multiple modules to have different set of
defaultable properties.

Remove unrelated `features()` method of `defaultable` interface.

Make `generateCommon` to be `defaultable` as it uses other defaults
by `flag_defaults` property.

Signed-off-by: Sebastian Birunt <sebastian.birunt@arm.com>
Change-Id: Iab922fab7eb72d0cf676aa6103e26d2e8a6a30e3